feat(repo): Migrate handshake tests to e2e test suite
