--- jest.config.handshake.js
@@ -1,19 +1,0 @@
-/** @type {import('ts-jest').JestConfigWithTsJest} */
-module.exports = {
-  extensionsToTreatAsEsm: ['.ts'],
-  testRegex: ['handshake.test.tsx?$'],
-  moduleNameMapper: {
-    '^(\\.{1,2}/.*)\\.js$': '$1',
-  },
-  transform: {
-    // '^.+\\.[tj]sx?$' to process js/ts with `ts-jest`
-    // '^.+\\.m?[tj]sx?$' to process js/ts/mjs/mts with `ts-jest`
-    '^.+\\.tsx?$': [
-      'ts-jest',
-      {
-        diagnostics: false,
-        useESM: true,
-      },
-    ],
-  },
-};
